fix cron-task delete job code

+++ internal/computing/cron_task.go
@@ -301,7 +301,6 @@ func (task *CronTask) watchExpiredTask() {
 		var deleteSpaceIdAndJobUuid = make(map[string]string)
 		for _, jobCopy := range jobList {
 			job := jobCopy
-			logs.GetLogger().Infof("debug:: job_uuid: %s, space_uuid: %s, status: %s", job.JobUuid, job.SpaceUuid, models.GetJobStatus(job.Status))
 			jobUuidDeployName := constants.K8S_DEPLOY_NAME_PREFIX + strings.ToLower(job.JobUuid)
 			if _, ok := deployOnK8s[jobUuidDeployName]; ok {
 				var nameSpace = job.NameSpace
@@ -318,7 +317,18 @@ func (task *CronTask) watchExpiredTask() {
 				// compatible with space_uuid
 				spaceUuidDeployName := constants.K8S_DEPLOY_NAME_PREFIX + strings.ToLower(job.SpaceUuid)
 				if _, ok = deployOnK8s[spaceUuidDeployName]; ok {
-					if NewJobService().GetJobEntityBySpaceUuid(job.SpaceUuid) > 0 {
+					if NewJobService().GetJobEntityBySpaceUuid(job.SpaceUuid) > 0 && time.Now().Unix() < job.ExpireTime {
+						if job.Status != models.JOB_RUNNING_STATUS {
+							foundDeployment, err := NewK8sService().k8sClient.AppsV1().Deployments(job.NameSpace).Get(context.TODO(), job.K8sDeployName, metav1.GetOptions{})
+							if err != nil {
+								continue
+							}
+							if foundDeployment.Status.AvailableReplicas > 0 {
+								job.PodStatus = models.POD_RUNNING_STATUS
+								job.Status = models.JOB_RUNNING_STATUS
+								NewJobService().UpdateJobEntityByJobUuid(job)
+							}
+						}
 						continue
 					}
 
@@ -336,8 +346,6 @@ func (task *CronTask) watchExpiredTask() {
 				}
 			}
 
-			logs.GetLogger().Infof("debug:: job_uuid: %s, 1111111", job.JobUuid)
-
 			if job.DeleteAt == models.DELETED_FLAG {
 				continue
 			}
@@ -346,17 +354,12 @@ func (task *CronTask) watchExpiredTask() {
 			createdTime := time.Unix(job.CreateTime, 0)
 			createDuration := currentTime.Sub(createdTime)
 
-			logs.GetLogger().Infof("debug:: job_uuid: %s createTime: %f, expiredTime: %t,  222222", job.JobUuid, createDuration.Hours(), time.Now().Unix() > job.ExpireTime)
-
-			if createDuration.Hours() <= 2 && job.Status != models.JOB_RUNNING_STATUS {
-				continue
-			}
-
-			logs.GetLogger().Infof("debug:: job_uuid: %s, 33333333", job.JobUuid)
-
 			if job.NameSpace != "" && job.K8sDeployName != "" {
 				foundDeployment, err := NewK8sService().k8sClient.AppsV1().Deployments(job.NameSpace).Get(context.TODO(), job.K8sDeployName, metav1.GetOptions{})
 				if err != nil {
+					if createDuration.Hours() <= 2 && job.Status != models.JOB_RUNNING_STATUS {
+						continue
+					}
 					if errors.IsNotFound(err) {
 						// delete job
 						logs.GetLogger().Warnf("not found deployment on the cluster, job_uuid: %s, deployment: %s", job.JobUuid, job.K8sDeployName)
@@ -367,7 +370,7 @@ func (task *CronTask) watchExpiredTask() {
 					continue
 				}
 
-				if foundDeployment.Status.AvailableReplicas == 0 { // need to delete
+				if foundDeployment.Status.AvailableReplicas == 0 && createDuration.Hours() > 2 { // need to delete
 					DeleteJob(job.NameSpace, job.JobUuid, "cron-task correction status")
 					deleteSpaceIdAndJobUuid[job.JobUuid] = job.SpaceUuid + "_" + job.JobUuid
 					continue
@@ -382,7 +385,6 @@ func (task *CronTask) watchExpiredTask() {
 
 			checkFcpJobInfoInChain(job)
 
-			logs.GetLogger().Infof("debug:: job_uuid: %s, 44444444", job.JobUuid)
 			if job.Status == models.JOB_TERMINATED_STATUS || job.Status == models.JOB_COMPLETED_STATUS || time.Now().Unix() > job.ExpireTime {
 				expireTime := time.Unix(job.ExpireTime, 0).Format("2006-01-02 15:04:05")
 				logs.GetLogger().Infof("job_uuid: %s, current status is %s, expire time: %s, starting to delete it.", job.JobUuid, models.GetJobStatus(job.Status), expireTime)
